Merged PR 123931: Fixing small details to be in the exact standard we have + Final Case for Brainstorm
Reviewed by Chris, Anna did not had time to do it. I created this "code-review" branch to test a review process we are planning (me, Anna and Chris).

Standard messages for contribution (collaboration) and also the structure/name of the files. Changed based on the Teams discussion. Also fixed a broken link in the introduction. Also, added a Final Case challenge for brainstorm as group activity, as the other bootcamps. I used the same use case of Anna's training, including the link for hers materials at GitHub.

Related work items: #108950